move translation controllers to own file
